Make CompilerError top level & put back_end in its own folder
CompilerError is a defined exception type that we use throughout the
compiler (all the modules that currently exist). It's been moved to
the top level where it belongs.

`back_end` is now in its own folder, where we'll be able to start
the process of transforming it from a monolith to individual modules.